update created_or_updated processor
